+++ src/BlockEntities/HopperEntity.h
@@ -78,7 +78,7 @@ protected:
 	bool MoveItemsFromGrid(cBlockEntityWithItems & a_Entity);
 
 	/** Moves one piece from the specified itemstack into this hopper. Returns true if contents have changed. Doesn't change the itemstack. */
-	bool MoveItemsFromSlot(cBlockEntityWithItems & a_Entity, int a_SrcSlotNum);
+	bool MoveItemsFromSlot(cBlockEntityWithItems & a_Entity, size_t a_SrcSlotNum);
 
 	/** Moves items to the chest at the specified absolute coords. Returns true if contents have changed */
 	bool MoveItemsToChest(cChunk & a_Chunk, Vector3i a_Coords);
